Extract code into `SyntaxRule` protocol
